getElementAttributes to getElementData
