+++ src/tools/html/index.ts
@@ -1,5 +1,5 @@
 export * from './find-links.js';
-export * from './get-element-attributes.js';
+export * from './get-element-data.js';
 export * from './get-page-data.js';
 export * from './get-plaintext.js';
 export * from './parse-with-cheerio.js';
